+++ R/adjacency_matrix.R
@@ -1,31 +1,41 @@
 #' Adjacency matrix
 #'
 #' Create adjacency matrix of HPO child-parent relationships.
-#' @param pheno_ids Character vector of HPO IDs.
-#' @param as_dataframe Can return matrix or dataframe \<bool\>
+#' @param terms Character vector of HPO IDs.
 #' @inheritParams make_phenos_dataframe
+#' @inheritParams ontologyIndex::get_term_descendancy_matrix
 #' @returns adjacency matrix
 #'
 #' @export
+#' @importFrom ontologyIndex get_term_descendancy_matrix
 #' @examples
-#' adjacency <- adjacency_matrix(pheno_ids = c("HP:000001", "HP:000002"))
-adjacency_matrix <- function(pheno_ids,
-                             hpo = get_hpo(),
-                             as_dataframe = FALSE) {
-    message("adjacency_matrix")
-    HPO_id <- unique(pheno_ids)
-    size <- length(HPO_id)
-    adjacency <- data.frame(matrix(nrow = size, ncol = size))
-    rownames(adjacency) <- HPO_id
-    colnames(adjacency) <- HPO_id
-    adjacency[is.na(adjacency)] <- 0
-    for (id in HPO_id) {
-        children <- hpo$children[id][[1]]
-        adjacency[id, children] <- 1
+#' adjacency <- adjacency_matrix(terms = c("HP:0000001", "HP:0000002"))
+adjacency_matrix <- function(hpo = get_hpo(),
+                             terms = unique(hpo$id),
+                             verbose = TRUE) {
+
+   messager("Creating adjacency matrix for",
+            formatC(length(terms),big.mark = ","),"terms.",
+            v=verbose)
+  #### Custom script #####
+  terms <- unique(terms)
+  terms <- terms[terms %in% hpo$id]
+  if(length(terms)==0) stop("0 valid HPO IDs (terms) provided.")
+  size <- length(terms)
+  adjacency <- matrix(nrow = size,
+                      ncol = size,
+                      dimnames = list(terms, terms))
+  adjacency[is.na(adjacency)] <- 0
+  for (id in terms) {
+    children <- hpo$children[[id]]
+    children <- children[children %in% terms]
+    if(length(children)>0){
+      adjacency[id, children] <- 1
     }
-    if (isTRUE(as_dataframe)) {
-        return(adjacency[HPO_id, HPO_id])
-    } else {
-        return(as.matrix(adjacency[HPO_id, HPO_id]))
-    }
+  }
+  #### ontologyIndex method ####
+  ## Returns same object but not sure if it's the same in all conditions
+  # adjacency <- ontologyIndex::get_term_descendancy_matrix(ontology = hpo,
+  #                                                         terms = terms) *1
+  return(adjacency)
 }